lambdas/image_processor/requirements.txt
boto3==1.38.1
botocore==1.38.1
dotenv==0.9.9
jmespath==1.0.1
pillow==11.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
s3transfer==0.12.0
six==1.17.0
urllib3==2.4.0
